+++ src/Configuration.php
@@ -2,8 +2,11 @@
 
 namespace h4kuna\Exchange;
 
+use h4kuna\Exchange\Driver\Cnb\Day;
+
 final class Configuration
 {
+
 	public function __construct(public string $from, public string $to)
 	{
 	}