+++ src/Driver/DriverAccessor.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace h4kuna\Exchange\Driver;
+
+interface DriverAccessor
+{
+
+	function get(string $name): Driver;
+
+}